+++ .gitignore
@@ -35,3 +35,5 @@ pom.xml.versionsBackup
 *.orig
 *.old
 *.md.html
+/.apt_generated/
+/.apt_generated_tests/
