+++ error_code_config.yml
@@ -2,3 +2,4 @@ error-tags:
   BFSVS:
     packages:
       - com.exasol.adapter.document
+    highest-index: 7
