+++ src/test/java/com/exasol/adapter/document/documentfetcher/files/AbstractLocalFileLoaderTest.java
@@ -45,8 +45,8 @@ class AbstractLocalFileLoaderTest {
         final FileLoaderStub fileLoader = new FileLoaderStub(tempDir, WildcardExpression.fromGlob("/testFile*.json"));
         final List<RemoteFile> result = new ArrayList<>();
         fileLoader.loadFiles().forEachRemaining(result::add);
-        final List<String> firstLines = readFirstLineFromStreams(
-                result.stream().map(RemoteFile::getInputStream).collect(Collectors.toList()));
+        final List<String> firstLines = readFirstLineFromStreams(result.stream()
+                .map(remoteFile -> remoteFile.getContent().getInputStream()).collect(Collectors.toList()));
         final List<String> resourceNames = result.stream().map(RemoteFile::getResourceName)
                 .collect(Collectors.toList());
         assertAll(//
